fix; Error message by incorrect aniworld url
